separate out BDO LP logic
